Control.Concurrent.RCU.Internal

Description

STM-based RCU with concurrent writers

Documentation

newtype SRef s a

Shared references

Constructors

SRef

Fields

unSRef :: TVar a

data RCUThread s a

This is a basic RCU thread. It may be embellished when running in a more exotic context.

Constructors

RCUThread

Fields

rcuThreadId :: !ThreadId

rcuThreadVar :: !(MVar a)

newtype RCU s a

This is an RCU computation. It can use forking and joining to form new threads, and then you can use reading and writing to run classic read-side and write-side RCU computations. Contention between multiple write-side computations is managed by STM.

Constructors

RCU

Fields

unRCU :: TVar Int64 -> IO a

runRCU :: (forall s. RCU s a) -> IO a

Run an RCU computation.

newtype ReadingRCU s a

This is the basic read-side critical section for an RCU computation

Constructors

ReadingRCU

Fields

runReadingRCU :: IO a

newtype WritingRCU s a

This is the basic write-side critical section for an RCU computation

Constructors

WritingRCU

Fields

runWritingRCU :: TVar Int64 -> STM a

class Monad m => MonadNew s m | m -> s where

Minimal complete definition

Nothing

Methods

newSRef :: a -> m (SRef s a)

Build a new shared reference

class MonadNew s m => MonadReading s m | m -> s where

This is a read-side critical section

Minimal complete definition

Nothing

Methods

readSRef :: SRef s a -> m a

Read a shared reference.

copySRef :: MonadReading s m => SRef s a -> m (SRef s a)

Copy a shared reference.

class MonadReading s m => MonadWriting s m | m -> s where

This is a write-side critical section

Minimal complete definition

Nothing

Methods

writeSRef :: SRef s a -> a -> m ()

Write to a shared reference.

synchronize :: m ()

Synchronize with other writers.

No other writer can straddle this time bound. It will either see writes from before, or writes after, but never some of both!

class (MonadReading s (Reading m), MonadWriting s (Writing m), MonadNew s m) => MonadRCU s m | m -> s where

This is the executor service that can fork, join and execute critical sections.

Associated Types

type Reading m :: * -> *

A read-side critical section

type Writing m :: * -> *

A write-side critical section

type Thread m :: * -> *

Threads we can fork and join

Methods

forking :: m a -> m (Thread m a)

Fork a thread

joining :: Thread m a -> m a

Join a thread

reading :: Reading m a -> m a

Run a read-side critical section

writing :: Writing m a -> m a

Run a write-side critical section